refactor: rename files
Remove redundant "service" from module names. "service" is already in the module-path.
